refactor: decompose hubChatActions.ts (3001 lines) into per-module handlers
Split the monolithic executeAction switch into:
- chatActions/types.ts — all action + storage interfaces (581 lines)
- chatActions/finykActions.ts — finance actions (450 lines, 15 cases)
- chatActions/fizrukActions.ts — fitness actions (700 lines, 12 cases)
- chatActions/routineActions.ts — habit actions (393 lines, 10 cases)
- chatActions/nutritionActions.ts — nutrition actions (398 lines, 10 cases)
- chatActions/crossActions.ts — cross-module/analytics/utils/memory (482 lines, 14 cases)
- hubChatActions.ts — thin router (24 lines)

Public API (executeAction, ChatAction) unchanged.
All 704 tests pass. Build unchanged.
